To make even funkier Fireballs, we need to kill da boss
The boss has HP now, which won't make sense since nothing can kill it here, but when it's HP drops to 0, the style will change and make all of the fireballs shake up and surround it before exploding outwards. The fireballs also have a bool that tells it weather it's alive that can be switched off. For testing purposes, the boss' HP goes down very slowly, and the fireballs randomly die.